// src/api/generated/models/hardSkillLevelsEnum.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Brew
 * OpenAPI spec version: 0.0.1
 */

export type HardSkillLevelsEnum = typeof HardSkillLevelsEnum[keyof typeof HardSkillLevelsEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const HardSkillLevelsEnum = {
  fundamentals: 'fundamentals',
  basic: 'basic',
  intermediate: 'intermediate',
  advanced: 'advanced',
  expert: 'expert',
  authority: 'authority',
} as const;