// src/api/generated/models/jobVacancyResource.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Brew
 * OpenAPI spec version: 0.0.1
 */
import type { JobVacancyResourceTitlePt } from './jobVacancyResourceTitlePt';
import type { JobVacancyResourceTitleEn } from './jobVacancyResourceTitleEn';
import type { JobVacancyResourceDescriptionPt } from './jobVacancyResourceDescriptionPt';
import type { JobVacancyResourceDescriptionEn } from './jobVacancyResourceDescriptionEn';
import type { EmploymentType } from './employmentType';
import type { JobVacancyResourceBenefitsPt } from './jobVacancyResourceBenefitsPt';
import type { JobVacancyResourceBenefitsEn } from './jobVacancyResourceBenefitsEn';
import type { JobVacancyResourceTranslationStatusLabel } from './jobVacancyResourceTranslationStatusLabel';
import type { ContractType } from './contractType';
import type { SeniorityLevelEnum } from './seniorityLevelEnum';
import type { JobVacancyResourceSpecialties } from './jobVacancyResourceSpecialties';
import type { JobVacancyStatusEnum } from './jobVacancyStatusEnum';
import type { JobVacancyResourceStatusLabel } from './jobVacancyResourceStatusLabel';
import type { SelectionProcessStageEnum } from './selectionProcessStageEnum';
import type { JobVacancyResourceProcessStepLabel } from './jobVacancyResourceProcessStepLabel';
import type { LanguageResource } from './languageResource';
import type { SoftSkillResource } from './softSkillResource';
import type { JobVacancyProcessStepResource } from './jobVacancyProcessStepResource';
import type { JobVacancyResourceCompanyProfileId } from './jobVacancyResourceCompanyProfileId';
import type { JobVacancyResourceProfile } from './jobVacancyResourceProfile';

export interface JobVacancyResource {
  id: string;
  title: string;
  title_pt: JobVacancyResourceTitlePt;
  title_en: JobVacancyResourceTitleEn;
  description: string;
  description_pt: JobVacancyResourceDescriptionPt;
  description_en: JobVacancyResourceDescriptionEn;
  employment_type: EmploymentType;
  benefits: unknown[];
  benefits_pt: JobVacancyResourceBenefitsPt;
  benefits_en: JobVacancyResourceBenefitsEn;
  translation_status: string;
  translation_status_label: JobVacancyResourceTranslationStatusLabel;
  estimated_salary: string;
  contract_type: ContractType;
  seniority_level: SeniorityLevelEnum;
  specialties: JobVacancyResourceSpecialties;
  status: JobVacancyStatusEnum;
  status_label: JobVacancyResourceStatusLabel;
  process_step: SelectionProcessStageEnum;
  process_step_label: JobVacancyResourceProcessStepLabel;
  languages?: LanguageResource[];
  soft_skills?: SoftSkillResource[];
  language_desirable?: LanguageResource[];
  process_steps?: JobVacancyProcessStepResource[];
  company_profile_id: JobVacancyResourceCompanyProfileId;
  profile?: JobVacancyResourceProfile;
  dev_profiles_count?: number;
}